refactor: вынести преобразование ответов от сторонней API  в отдельный файл
