+++ src/pages/search-page/search-page.module.css
@@ -1,0 +1,6 @@
+.text {
+  color: var(--light-theme-black-text-color);
+}
+.textDark {
+  color: var(--dark-theme-title-color);
+}
